split react EndpointAggregationForm from react BaseVisForm
